Move strategies out of the main js file. Rename few names
